feat: make auto in-air mantling optional
The default is preserved for backwards compatibility (true by default).
When disabled, then player/ai input is required in order to trigger
mantling while in air, just as it is while on the ground.

The example has also been updated to detect the state of this setting.
If disabled, and the character is in air, then will attempt to perform
in-air mantling as part of the OnJump input handler.